+++ CHANGELOG.md
@@ -2,6 +2,11 @@ CHANGELOG
 =========
 
 * version actuelle (non versionné pour le moment) :
+  * Test en cours
+
+----------------------------------------------------
+
+* 2023032500 :
   * Klipper : 
     * Configuration du /etc/default/klipper
     * Ajout du fichier "variables" pour le stockage 
@@ -11,7 +16,8 @@ CHANGELOG
     * Configuration de l'API KEY et du thème
     * Configuration des interpréteurs : xinit + bash + zsh
     * Autologin
-
-  
+  * Installation :
+    * wiringpi via le code source
+    * nodejs
 * 2023031600 :
   * Réécriture des scripts from scratch